[lldb-dap] generate lldb launch commands for android APKs
