+++ lldb/tools/lldb-dap/src-ts/android/apk-debug-configuration.ts
@@ -1,13 +1,18 @@
-import * as os from "node:os";
 
 export class ApkDebugConfiguration {
 
-    static getLldbLaunchCommands(deviceId: string | undefined, componentName: string): string[] {
-        // TODO: return the real commands needed to connect to lldb-server; last-start-commands should not be used here
-        // TODO: manage deviceId
-        const home = os.homedir();
+    static getLldbLaunchCommands(deviceSerial: string | undefined, componentName: string): string[] {
+        if (deviceSerial === undefined) {
+            deviceSerial = "";
+        }
+        const appId = componentName.split("/")[0];
         return [
-            `command source -s 0 -e 0 '${home}/.lldb/android/last-start-commands'`
+            `platform select remote-android`,
+            `platform connect unix-abstract-connect://${deviceSerial}/${appId}/lldb-platform.sock`,
+            `process attach --name ${appId}`,
+            `process handle SIGSEGV -n false -p true -s false`,
+            `process handle SIGBUS -n false -p true -s false`,
+            `process handle SIGCHLD -n false -p true -s false`,
         ];
     }
 }
